Move SingleEvent file and TODO on Checkboxes.
